requirements.txt
Flask==3.0.3
PyYAML==6.0.2
matplotlib==3.10.8
numpy==2.2.6
opencv_python==4.12.0.88
pandas==3.0.0
Pillow==12.1.0
scikit_learn==1.8.0
scipy==1.17.0
Shapely==2.1.2
scikit-image==0.26.0
trimesh==4.10.1
mapbox_earcut==2.0.0
lxml==6.0.2
pywebview==5.1.1
nuitka==2.4.9
